Cleanup the code
Cleans up code to be a little bit more idiomatic, still a work in
progress
