+++ README.md
@@ -1,2 +1,5 @@
 # ReactNativeTodoApp
+
 react-native forgerock todo app
+
+This project works with xcode 12
